requirements.txt
SQLAlchemy~=1.4.15
mysql-connector-python
python-dotenv~=0.17.1
